Moved logger from common/models to common
